Add "next_available" and "available_until" values to rooms
When the room isn't available, but will become available
later on in the day, this will be a timestamp of the
next available slot.

When the room is currently available and there's a
subsequent event today, this will be a timestamp of the
start of the next event in the room.